refactor: move component files
