+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
 import PageContainer from "@/components/layout/page-container";
-import PageFooter from "@/components/layout/page-footer/page-footer";
-import PageHeader from "@/components/layout/page-header/page-header";
+import PageFooter from "@/components/layout/page-footer";
+import PageHeader from "@/components/layout/page-header";
 import ReactQueryProvider from "@/components/providers/react-query";
 import { StoreProvider } from "@/components/providers/zustand";
 import { Toaster } from "@/components/ui/toaster";
